Replace kol-pagination with kol-pagination-wc in table / Add kol-pagination-wc and  remove Host wrapper
Refs: #7126
