+++ packages/components/src/components/pagination/style.scss
@@ -1,29 +1,6 @@
 @use '../@shared/mixins' as *;
 @use '../../styles/global' as *;
 @use '../tooltip/style.scss' as *;
+@use '../@shared/kol-pagination-mixin' as *;
 
-@layer kol-component {
-	/* TODO. Once #7126 is completed, move :host block into .kol-pagination  */
-	:host {
-		align-items: center;
-		display: grid;
-		gap: rem(16);
-		grid-template-columns: 1fr auto;
-	}
-
-	.kol-pagination {
-		&__navigation-list {
-			align-items: center;
-			display: inline-flex;
-			flex-wrap: wrap;
-			gap: 0.5em;
-			list-style: none;
-			margin: 0;
-			padding: 0;
-		}
-
-		&__separator:before {
-			content: '•••';
-		}
-	}
-}
+@include kol-pagination-styles;
